Refactored get_site_id from cms.utils.i18n -> cms.utils.conf
