+++ cms/utils/__init__.py
@@ -1,8 +1,8 @@
 # -*- coding: utf-8 -*-
 # TODO: this is just stuff from utils.py - should be splitted / moved
 from cms import constants
-from cms.utils.conf import get_cms_setting
-from cms.utils.i18n import get_default_language, get_language_list, get_language_code, get_site_id
+from cms.utils.conf import get_cms_setting, get_site_id
+from cms.utils.i18n import get_default_language, get_language_list, get_language_code
 from django.conf import settings
 from django.core.files.storage import get_storage_class
 from django.utils.functional import LazyObject
